[#42] Removed unnecessary comments and lines
