+++ src/test/java/org/example/Operations/DataObjectOperationsTest.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import org.example.Mapper.Serialize.ModifyMetadataOperations;
-import org.example.Mapper.Serialize.ModifyPermissionsOperations;
+import org.example.Serialize.ModifyMetadataOperations;
+import org.example.Serialize.ModifyPermissionsOperations;
 import org.example.Properties.DataObject.*;
 import org.example.Properties.Resource.ResourceCreateParams;
 import org.example.Util.Response;
